refactor: extract toml_parser tests to separate file
Moves tests to toml_parser_tests.rs to keep toml_parser.rs under the
1000-line readability ratchet threshold.
